Finish moving from `/docs/docs` to `/docs`
Added local hosting instructions in the YAML file

This renders beautifully locally, but for some reason it doesn't render properly on GitHub Pages yet. One possible reason is the `pages-build-deployment` workflow that GitHub Pages created on my initial attempt to use Pages (before trying MkDocs) is still running (and I can't figure out how to get rid of it) and it appears to be interfering with the `docs` workflow I created for MkDocs

If someone has ideas on how to debug this, please open an issue or email me - thx!